requirements.txt
# requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn



